Update Sum.cpp
The input starts with a single integer, N, in a single line. N denotes the total number of schools. N lines will then follow with 2 integers each, the i-th line will contain Xi and Yi. This indicates the i-th school sent Xi students for the secondary division and Yi students for the junior college division.
Your program should print two integers on a line separated by a space. The first integer is the total number of students in the secondary division and the second integer is the total number of students in the junior college division.
-> maximum value of n is 10000000 and minimum value of n is 2
2
1 4
2 3 // out put : 3 7